# docs/swagger.yaml
definitions:
  APIError:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  CreateNewsletterReq:
    properties:
      description:
        type: string
      name:
        type: string
      picture:
        type: string
    required:
    - name
    type: object
...
  NewsletterMarkViewedReq:
    properties:
      newsletterJid:
        type: string
      serverIds:
        items:
          type: integer
        type: array
    required:
    - newsletterJid
    - serverIds
    type: object
  NewsletterMessageReq:
    properties:
      beforeId:
        type: integer
      count:
        type: integer
      newsletterJid:
        type: string
    required:
    - newsletterJid
    type: object
  NewsletterMuteReq:
    properties:
      mute:
        type: boolean
      newsletterJid:
        type: string
    required:
    - newsletterJid
    type: object
  NewsletterReactReq:
    properties:
      messageId:
        type: string
      newsletterJid:
        type: string
      reaction:
        type: string
      serverId:
        type: integer
    required:
    - messageId
    - newsletterJid
    - reaction
    - serverId
    type: object
  NewsletterSubscribeReq:
    properties:
      newsletterJid:
        type: string
    required:
    - newsletterJid
    type: object
...
info:
  contact:
    name: wzap Support
    url: https://github.com/felipyfgs/wzap
  description: WhatsApp Multi-Session API powered by whatsmeow
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: wzap API
  version: "1.0"
paths:
...
  /sessions/{sessionId}/newsletter/create:
    post:
      consumes:
      - application/json
      description: Creates a new WhatsApp Newsletter (channel) with optional description
        and profile picture
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Newsletter payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateNewsletterReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Create a newsletter
      tags:
      - Newsletter
  /sessions/{sessionId}/newsletter/info:
    post:
      description: Retrieves metadata about a newsletter by its JID
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Newsletter JID
        in: query
        name: jid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Get newsletter info
      tags:
      - Newsletter
  /sessions/{sessionId}/newsletter/invite:
    post:
      description: Retrieves newsletter metadata from an invite code without subscribing
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Newsletter invite code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Get newsletter info from invite code
      tags:
      - Newsletter
  /sessions/{sessionId}/newsletter/list:
    get:
      description: Returns all newsletters the current session is subscribed to
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: List subscribed newsletters
      tags:
      - Newsletter
  /sessions/{sessionId}/newsletter/messages:
    post:
      consumes:
      - application/json
      description: Fetches messages from a newsletter; supports pagination via before_id
        cursor and count
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Messages pagination payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/NewsletterMessageReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Get newsletter messages
      tags:
      - Newsletter
  /sessions/{sessionId}/newsletter/mute:
    post:
      consumes:
      - application/json
      description: Toggles mute status for a newsletter
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Mute payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/NewsletterMuteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Mute/unmute a newsletter
      tags:
      - Newsletter
  /sessions/{sessionId}/newsletter/react:
    post:
      consumes:
      - application/json
      description: Sends a reaction to a newsletter message
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Reaction payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/NewsletterReactReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: React to a newsletter message
      tags:
      - Newsletter
  /sessions/{sessionId}/newsletter/subscribe:
    post:
      consumes:
      - application/json
      description: Subscribes the current session to a newsletter identified by its
        JID
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Newsletter JID payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/NewsletterSubscribeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Subscribe to a newsletter
      tags:
      - Newsletter
  /sessions/{sessionId}/newsletter/unsubscribe:
    post:
      consumes:
      - application/json
      description: Unfollows a newsletter
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Newsletter JID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/NewsletterSubscribeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Unsubscribe from a newsletter
      tags:
      - Newsletter
  /sessions/{sessionId}/newsletter/viewed:
    post:
      consumes:
      - application/json
      description: Marks newsletter messages as viewed by server IDs
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Mark viewed payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/NewsletterMarkViewedReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Mark newsletter messages as viewed
      tags:
      - Newsletter
...
securityDefinitions:
  Authorization:
    description: Session or admin token (without Bearer prefix)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"